learner/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
pydantic[email]==2.11.9
pydantic-settings==2.11.0
python-dotenv==1.1.1